Move shouldUseOneRevisionUpdateToSplitFilesInDifferentPartitions
