+++ java/statestore/src/test/java/sleeper/statestore/s3/S3StateStoreNewIT.java
@@ -16,10 +16,14 @@
 
 package sleeper.statestore.s3;
 
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
 
 import sleeper.core.schema.type.LongType;
 import sleeper.core.statestore.FileReference;
+import sleeper.core.statestore.SplitFileReferences;
 
 import java.time.Instant;
 import java.util.HashSet;
@@ -32,21 +36,56 @@ import static sleeper.core.schema.SchemaTestHelper.schemaWithKey;
 
 public class S3StateStoreNewIT extends S3StateStoreNewTestBase {
 
-    @Test
-    public void shouldAddAndRetrieve1000FileReferences() throws Exception {
-        // Given
-        initialiseWithSchema(schemaWithKey("key", new LongType()));
-        List<FileReference> files = IntStream.range(0, 1000)
-                .mapToObj(i -> factory.rootFile("file-" + i, 1))
-                .collect(Collectors.toUnmodifiableList());
-        store.fixTime(Instant.ofEpochMilli(1_000_000L));
+    @Nested
+    @DisplayName("Handle S3-specific cases for file references")
+    class FileReferences {
+        @BeforeEach
+        void setUp() throws Exception {
+            initialiseWithSchema(schemaWithKey("key", new LongType()));
+        }
 
-        // When
-        store.addFiles(files);
+        @Test
+        public void shouldAddAndRetrieve1000FileReferences() throws Exception {
+            // Given
+            List<FileReference> files = IntStream.range(0, 1000)
+                    .mapToObj(i -> factory.rootFile("file-" + i, 1))
+                    .collect(Collectors.toUnmodifiableList());
+            store.fixTime(Instant.ofEpochMilli(1_000_000L));
 
-        // Then
-        assertThat(new HashSet<>(store.getFileReferences())).isEqualTo(files.stream()
-                .map(reference -> withLastUpdate(Instant.ofEpochMilli(1_000_000L), reference))
-                .collect(Collectors.toSet()));
+            // When
+            store.addFiles(files);
+
+            // Then
+            assertThat(new HashSet<>(store.getFileReferences())).isEqualTo(files.stream()
+                    .map(reference -> withLastUpdate(Instant.ofEpochMilli(1_000_000L), reference))
+                    .collect(Collectors.toSet()));
+        }
+
+        @Test
+        void shouldUseOneRevisionUpdateToSplitFilesInDifferentPartitions() throws Exception {
+            // Given
+            initialiseWithSchema(schemaWithKey("key", new LongType()));
+            splitPartition("root", "L", "R", 5);
+            splitPartition("L", "LL", "LR", 2);
+            splitPartition("R", "RL", "RR", 7);
+            FileReference file1 = factory.partitionFile("L", "file1", 100L);
+            FileReference file2 = factory.partitionFile("R", "file2", 200L);
+            store.addFiles(List.of(file1, file2));
+
+            // When
+            SplitFileReferences.from(store).split();
+
+            // Then
+            assertThat(getCurrentFilesRevision()).isEqualTo(versionWithPrefix("3"));
+        }
+    }
+
+    private String getCurrentFilesRevision() {
+        S3RevisionIdStore revisionStore = new S3RevisionIdStore(dynamoDBClient, instanceProperties, tableProperties);
+        return revisionStore.getCurrentFilesRevisionId().getRevision();
+    }
+
+    private static String versionWithPrefix(String version) {
+        return "00000000000" + version;
     }
 }
